created PeristMenuData struct and added selectedArrayToUserInput func
